feat(bundler): enable dev bundler through flag
