+++ package.json
@@ -34,7 +34,7 @@
     "cross-env": "^7.0.2",
     "enzyme": "^3.11.0",
     "enzyme-to-json": "^3.2.2",
-    "eslint": "^7.32.0",
+    "eslint": "8",
     "eslint-config-amex": "^15.2.1",
     "eslint-plugin-jest": "^25.3.4",
     "eslint-plugin-jest-dom": "^4.0.1",
