+++ packages/one-app-dev-bundler/esbuild/plugins/one-app-index-loader-injectors/app-compatibility-injector.js
@@ -34,5 +34,5 @@ export default class AppCompatibilityInjector {
     appCompatibility: "${this.appCompatibility}",
   });
 `;
-  }
+  };
 }
