+++ packages/one-app-dev-bundler/esbuild/plugins/one-app-index-loader-injectors/holocron-module-register-injector.js
@@ -30,5 +30,5 @@ export default class HolocronModuleRegisterInjector {
     return `${content};
 Holocron.registerModule("${this.moduleName}", ${rootComponentName});
 `;
-  }
+  };
 }
