+++ packages/one-app-dev-bundler/esbuild/plugins/one-app-index-loader-injectors/module-metadata-injector.js
@@ -35,5 +35,5 @@ ${rootComponentName}.${META_DATA_KEY} = {
   version: '${this.version}',
 };
 `;
-  }
+  };
 }
