+++ packages/one-app-dev-bundler/esbuild/plugins/one-app-index-loader-injectors/provided-externals-injector.js
@@ -60,5 +60,5 @@ ${this.globalReferenceString}.getTenantRootModule = () => ${rootComponentName};
 ${this.globalReferenceString}.rootModuleName = '${this.moduleName}';
 }
 `;
-  }
+  };
 }
